requirements.txt
asgiref==3.11.0
cssselect==1.3.0
d==0.2.2
Django==6.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
lxml==6.0.2
Markdown==3.10
Pygments==2.19.2
PyJWT==2.10.1
pyquery==2.0.1
sqlparse==0.5.5
